requirements.txt
# Core AI + LangChain
langchain>=0.1.20
langchain-groq>=0.1.3

# Groq LLMs
groq>=0.5.0

# Embeddings
sentence-transformers>=2.2.2
huggingface-hub>=0.20.3

# Vector Store
chromadb>=0.4.24

# Utilities
tqdm
pydantic>=2.0
python-dotenv
typing-extensions>=4.6.1

# For parsing and logging
astunparse
